+++ podcasts/PodcastListViewController+Tip.swift
@@ -34,19 +34,11 @@ extension PodcastListViewController: UIPopoverPresentationControllerDelegate {
         vc.view.backgroundColor = .clear
         vc.view.clipsToBounds = false
         vc.modalPresentationStyle = .popover
-        if #available(iOS 16.0, *) {
-            vc.sizingOptions = [.preferredContentSize]
-        } else {
-            vc.preferredContentSize = idealSize
-        }
+        vc.sizingOptions = [.preferredContentSize]
         if let popoverPresentationController = vc.popoverPresentationController {
             popoverPresentationController.delegate = self
             popoverPresentationController.permittedArrowDirections = [.up]
-            if #available(iOS 16.0, *) {
-                popoverPresentationController.sourceItem = button
-            } else {
-                popoverPresentationController.barButtonItem = button
-            }
+            popoverPresentationController.sourceItem = button
             popoverPresentationController.backgroundColor = ThemeColor.primaryUi01()
         }
         return vc