+++ podcasts/PodcastViewController.swift
@@ -1280,11 +1280,7 @@ class PodcastViewController: FakeNavViewController, PodcastActionsDelegate, Sync
         vc.view.backgroundColor = .clear
         vc.view.clipsToBounds = false
         vc.modalPresentationStyle = .popover
-        if #available(iOS 16.0, *) {
-            vc.sizingOptions = [.preferredContentSize]
-        } else {
-            vc.preferredContentSize = idealSize
-        }
+        vc.sizingOptions = [.preferredContentSize]
         if let popoverPresentationController = vc.popoverPresentationController {
             popoverPresentationController.delegate = self
             popoverPresentationController.permittedArrowDirections = [.down]
@@ -1345,11 +1341,7 @@ class PodcastViewController: FakeNavViewController, PodcastActionsDelegate, Sync
         vc.view.backgroundColor = .clear
         vc.view.clipsToBounds = false
         vc.modalPresentationStyle = .popover
-        if #available(iOS 16.0, *) {
-            vc.sizingOptions = [.preferredContentSize]
-        } else {
-            vc.preferredContentSize = idealSize
-        }
+        vc.sizingOptions = [.preferredContentSize]
         if let popoverPresentationController = vc.popoverPresentationController {
             popoverPresentationController.delegate = self
             popoverPresentationController.permittedArrowDirections = [.down]