+++ podcasts/ProfileViewController.swift
@@ -11,7 +11,11 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
 
     @IBOutlet var footerView: UIView!
     @IBOutlet var alertIcon: UIImageView!
-    @IBOutlet var lastRefreshTime: UILabel!
+    @IBOutlet var lastRefreshTime: UILabel! {
+        didSet {
+            lastRefreshTime.font = UIFont.font(with: .body)
+        }
+    }
     @IBOutlet var refreshBtn: AnimatedImageButton! {
         didSet {
             refreshBtn.mainColor = ThemeColor.primaryText02()
@@ -138,7 +142,16 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
         }
 
         whatsNewDismissed()
-        showReferralsHintIfNeeded()
+
+        if FeatureFlag.cancelSubscriptionSurvey.enabled,
+           SyncManager.isUserLoggedIn(),
+           SubscriptionHelper.hasCancelledSubscription,
+           !Settings.subscriptionCancelledSurveyShown {
+            let controller = CancelSubscriptionSurveyViewModel.make()
+            present(controller, animated: true)
+        } else {
+            showReferralsHintIfNeeded()
+        }
     }
 
     override func viewDidDisappear(_ animated: Bool) {
@@ -302,6 +315,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
 
         cell.settingsImage.tintColor = ThemeColor.primaryIcon01()
         cell.settingsLabel.setLetterSpacing(-0.01)
+        cell.updateImageScale()
         cell.separatorInset = .zero
 
         switch row {
@@ -354,6 +368,16 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
         }
     }
 
+    func tableView(_ tableView: UITableView, estimatedHeightForRowAt indexPath: IndexPath) -> CGFloat {
+        let row = tableData[indexPath.section][indexPath.row]
+        switch row {
+        case .informationalBanner:
+            return 160
+        default:
+            return 70
+        }
+    }
+
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         let row = tableData[indexPath.section][indexPath.row]
 
@@ -363,7 +387,7 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
         } else if row == .informationalBanner {
             return 160
         } else {
-            return 70
+            return UITableView.automaticDimension
         }
     }