+++ podcasts/SwiftUI/EmptyStateView.swift
@@ -3,8 +3,34 @@ import SwiftUI
 protocol EmptyStateViewStyle: ObservableObject {
     var title: Color { get }
     var message: Color { get }
-    var button: Color { get }
     var icon: Color { get }
+    var button: Color { get }
+    var background: Color { get }
+}
+
+struct EmptyStateAction: Identifiable {
+    let id: String
+    let view: AnyView
+
+    init(
+        title: String,
+        action: @escaping () -> Void
+    ) {
+        self.id = title
+        self.view = AnyView(
+            Button(title) {
+                action()
+            }.buttonStyle(RoundedButtonStyle(theme: .sharedTheme))
+        )
+    }
+
+    init<Content: View>(
+        id: String,
+        @ViewBuilder content: () -> Content
+    ) {
+        self.id = id
+        self.view = AnyView(content())
+    }
 }
 
 /// Displays an informative view when there are no items to display and can be customized to show a custom view instead
@@ -12,6 +38,9 @@ protocol EmptyStateViewStyle: ObservableObject {
 ///
 /// The colors can be customized using EmptyStateViewStyle
 struct EmptyStateView<Title: View, Style: EmptyStateViewStyle>: View {
+    @ScaledMetric(relativeTo: .headline)
+   private var iconSize = 30
+
     @ObservedObject var style: Style
     let icon: (() -> Image)?
     let title: () -> Title
@@ -39,42 +68,34 @@ struct EmptyStateView<Title: View, Style: EmptyStateViewStyle>: View {
                 icon()
                     .resizable()
                     .foregroundStyle(style.icon)
-                    .frame(width: 30, height: 30)
+                    .frame(width: iconSize, height: iconSize)
             }
 
             title()
-                .font(style: .subheadline, weight: .semibold)
+                .font(.headline)
                 .foregroundStyle(style.title)
 
             if let message {
                 Text(message)
                     .multilineTextAlignment(.center)
-                    .font(style: .body)
+                    .font(.subheadline)
                     .foregroundStyle(style.message)
             }
 
-            HStack {
+            VStack {
                 ForEach(actions) { action in
-                    Button(action.title) {
-                        action.action()
-                    }.buttonStyle(RoundedButtonStyle(theme: .sharedTheme))
+                    action.view
                 }
             }
             .font(style: .subheadline, weight: .medium)
             .foregroundStyle(style.button)
         }
-        .frame(maxWidth: .infinity)
+        .frame(maxWidth: 400)
         .padding(.horizontal, EmptyConstants.padding)
         .padding(.vertical, EmptyConstants.verticalPadding)
         .padding(EmptyConstants.padding)
+        .background(style.background)
     }
-}
-
-struct EmptyStateAction: Identifiable {
-    let title: String
-    let action: () -> Void
-
-    var id: String { title }
 }
 
 private enum EmptyConstants {
@@ -85,14 +106,14 @@ private enum EmptyConstants {
 }
 
 extension EmptyStateView where Title == Text {
-    init(title: String, message: String?, icon: (() -> Image)? = nil, actions: [EmptyStateAction], style: Style) {
+    init(title: String, message: String?, icon: (() -> Image)? = nil, actions: [EmptyStateAction] = [], style: Style = .defaultStyle) {
         self.message = message
         self.actions = actions
+        self.icon = icon
         self.title = {
             Text(title)
         }
         self.style = style
-        self.icon = icon
     }
 }
 
@@ -109,7 +130,7 @@ struct EmptyStateView_Previews: PreviewProvider {
         var background: Color { .black }
         var title: Color { .white }
         var message: Color { .white.opacity(0.8) }
-        var button: Color { .red }
         var icon: Color { .primary }
+        var button: Color { .red }
     }
 }